Split apart XPCEncoder into smaller files
