+++ Sources/SecureXPC/XPC Coders/XPCDecoder/XPCDecoderImpl.swift
@@ -1,0 +1,34 @@
+//
+//  XPCDecoderImpl.swift
+//  
+//
+//  Created by Alexander Momchilov on 2021-11-12.
+//
+
+import Foundation
+
+fileprivate class XPCDecoderImpl: Decoder {
+	var codingPath = [CodingKey]()
+
+	let userInfo = [CodingUserInfoKey : Any]()
+
+	private let value: xpc_object_t
+
+
+	init(value: xpc_object_t, codingPath: [CodingKey]) {
+		self.value = value
+		self.codingPath = codingPath
+	}
+
+	func container<Key>(keyedBy type: Key.Type) throws -> KeyedDecodingContainer<Key> where Key : CodingKey {
+		return KeyedDecodingContainer(try XPCKeyedDecodingContainer(value: self.value, codingPath: self.codingPath))
+	}
+
+	func unkeyedContainer() throws -> UnkeyedDecodingContainer {
+		return try XPCUnkeyedDecodingContainer(value: self.value, codingPath: self.codingPath)
+	}
+
+	func singleValueContainer() throws -> SingleValueDecodingContainer {
+		return XPCSingleValueDecodingContainer(value: self.value, codingPath: self.codingPath)
+	}
+}
